+++ nodejs-learning/mongodb-crud/controller/userController.js
@@ -1,7 +1,5 @@
 const User = require("../models/userModel");
-const { hashPassword } = require("../models/userPassword");
 const httpStatus = require("http-status");
-const Joi = require("joi");
 
 const getAllUser = async (req, res) => {
   try {
@@ -20,117 +18,6 @@ const getAllUser = async (req, res) => {
       error: "Internal Server Error",
     });
   }
-};
-
-// CREATE NEW USER
-const userValidator = (req, res, next) => {
-  const schema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().email().required(),
-    password: Joi.string().min(6).max(10).required(),
-    password_confirmation: Joi.valid(Joi.ref("password")).required(),
-    tc: Joi.boolean().required(),
-  });
-
-  const { error } = schema.validate(req.body);
-  if (error) {
-    return res.status(httpStatus.BAD_REQUEST).json({
-      statusCode: httpStatus.BAD_REQUEST,
-      message: error.details[0].message,
-    });
-  }
-
-  next();
-};
-const createUser = async (req, res) => {
-  try {
-    const { name, email, password, password_confirmation, tc } = req.body;
-    const userExists = await User.getUserByEmail(email);
-    if (userExists) {
-      return res.status(httpStatus.CONFLICT).json({
-        statusCode: httpStatus.CONFLICT,
-        message: "Email already exists",
-      });
-    }
-    if (password !== password_confirmation) {
-      return res.status(httpStatus.BAD_REQUEST).json({
-        statusCode: httpStatus.BAD_REQUEST,
-        message: "Password and Confirm Password must match",
-      });
-    }
-    const hashedPassword = await hashPassword(password);
-    const savedUser = await User.createUser(name, email, hashedPassword, tc);
-    const saved_user = await User.getUserByEmail(email);
-    const token = User.generateToken(saved_user);
-
-    res.status(httpStatus.OK).json({
-      statusCode: httpStatus.OK,
-      message: "User create Success",
-      token: token,
-      savedUser: savedUser,
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(httpStatus.INTERNAL_SERVER_ERROR).json({
-      statusCode: httpStatus.INTERNAL_SERVER_ERROR,
-      message: "Unable to Register",
-    });
-  const { name, email, password, password_confirmation, tc } = req.body;
-  const userExists = await User.getUserByEmail(email);
-  if (userExists) {
-    res
-      .status(httpStatus.CONFLICT)
-      .json({
-        statusCode: `${httpStatus.CONFLICT}`,
-        message: "Email already exists",
-      });
-  } else {
-    if (name && email && password && password_confirmation && tc) {
-      if (password === password_confirmation) {
-        try {
-          const hashedPassword = await hashPassword(password);
-          const savedUser = await User.createUser(
-            name,
-            email,
-            hashedPassword,
-            tc
-          );
-          const saved_user = await User.getUserByEmail(email);
-          const token = User.generateToken(saved_user);
-          res
-            .status(httpStatus.OK)
-            .json({
-              statusCode: `${httpStatus.OK}`,
-              message: "User create Success",
-              token: token,
-              savedUser: savedUser,
-            });
-        } catch (error) {
-          res
-            .status(httpStatus.INTERNAL_SERVER_ERROR)
-            .json({
-              statusCode: `${httpStatus.INTERNAL_SERVER_ERROR}`,
-              message: "Unable to Register",
-            });
-        }
-      } else {
-        res
-          .status(httpStatus.NOT_FOUND)
-          .json({
-            statusCode: `${httpStatus.NOT_FOUND}`,
-            message: "Password and Confirm Password doesn't match",
-          });
-      }
-    } else {
-      res
-        .status(httpStatus.NOT_FOUND)
-        .json({
-          statusCode: `${httpStatus.NOT_FOUND}`,
-          message: "All fields are required",
-        });
-    }
-  }
-}
 };
 
 // GET USER BYID
@@ -197,9 +84,7 @@ const deleteUserById = async (req, res) => {
 
 module.exports = {
   getAllUser,
-  createUser,
   getUserById,
   updateUserById,
   deleteUserById,
-  userValidator,
 };
