+++ nodejs-learning/mongodb-crud/models/userSchema.js
@@ -3,7 +3,6 @@ const mongoose = require("mongoose");
 const userSchema = new mongoose.Schema({
   name: { type: String, require: true, trim: true },
   email: { type: String, require: true, trim: true, lowercase: true },
-  email: { type: String, require: true, trim: true },
   password: { type: String, require: true, trim: true },
   tc: { type: String, require: true, trim: true },
 });
