+++ nodejs-learning/mongodb-crud/package.json
@@ -17,7 +17,6 @@
     "http-status": "^1.6.2",
     "joi": "^17.10.0",
     "jsonwebtoken": "^9.0.1",
-    "mongoose": "^7.4.2",
-    "mysql2": "^3.6.0"
+    "mongoose": "^7.4.2"
   }
 }
