+++ nodejs-learning/mongodb-crud/routes/userRoutes.js
@@ -1,15 +1,22 @@
 const express = require("express");
 const router = express.Router();
 const studentController = require("../controller/userController");
-const loginController = require("../controller/loginController");
+const loginController = require("../controller/register-login-Controller");
+const { checkUserAuth } = require("../middleware/authMiddleware");
+
+router.use("/logged", checkUserAuth);
 
 router.post(
   "/create",
-  studentController.userValidator,
-  studentController.createUser
+  loginController.userValidator,
+  loginController.registerUser
 );
-router.post("/create", studentController.createUser);
-router.post("/login", loginController.loginUser);
+router.post(
+  "/login",
+  loginController.loginUserValidator,
+  loginController.loginUser
+);
+router.get("/logged", loginController.loggedUser);
 
 router.get("/", studentController.getAllUser);
 router.get("/:userId", studentController.getUserById);
